primefield: extract `dev` module
Moves all of the `test_*!` macros to the `dev` module
